Extract some of locationChange hook to separate action addCurrentPageToSummaryFilters
