+++ src/app/content/highlights/hooks/index.ts
@@ -1,5 +1,6 @@
 import { receiveUser } from '../../../auth/actions';
 import { actionHook } from '../../../utils';
+import addCurrentPageToSummaryFilters from './addCurrentPageToSummaryFilters';
 import createHighlight from './createHighlight';
 import filtersChange from './filtersChange';
 import loadHighlights from './locationChange';
@@ -9,6 +10,7 @@ import updateHighlight from './updateHighlight';
 export { loadHighlights };
 
 export default [
+  addCurrentPageToSummaryFilters,
   createHighlight,
   removeHighlight,
   updateHighlight,
